Codechange: Remove double chest from function parameters
Previously, we relied on duplicated code to copy values to double chests. Doors, however, were not applied too. To fix both these issues, I have added the applyToOtherContainer function that should be called after any modification of a block's NBT, so that any relevant value gets copied over to the other half of the chest or door.