Codechange: Refactor PluginIntegration friend-filtering API
Previously, we had to take in a full ArrayList<OfflinePlayer> into filterFriends which would be
modified using removeIf. This restricted us from being able to use something like ArrayList<UUID>
or even just integrating it into our own Stream. This is why I have decided to change the API so
that the Stream#filter calls or List#removeIf calls are not done by the PluginIntegration class
and are instead done by the caller. This allows for a more dynamic interface that can be adapted
to many more use-cases.