Improve performance of friend search & manually cache player profiles
The plugin now keeps a sqlite database for every UUID and Username pair. The friend search function now fully runs on a separate thread, and there is now also a "Loading..." screen until the first results are available, especially useful for larger servers.

This also switches from the zeroed UUID for representing the "public" to an invalid UUID, since more recent versions of Minecraft don't allow the zeroed UUID anymore. This is therefore a breaking change.